+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -101,6 +101,7 @@ our @EXPORT_OK = qw(
     isf_public_interface_schedule_report_temporal_contract_overlap_policy_values
     isf_public_interface_schedule_report_temporal_contract_reset_policy_shape
     isf_public_interface_schedule_report_top_level_keys
+    isf_public_interface_schedule_report_transaction_loop_keys
     isf_public_interface_schedule_report_transaction_wait_keys
     isf_public_interface_schedule_report_transaction_port_binding_keys
     isf_public_interface_schedule_report_transaction_port_binding_site_kind_values
@@ -244,6 +245,7 @@ sub build_isf_public_interface_contract {
         schedule_report_transaction_states_shape => isf_public_interface_schedule_report_transaction_states_shape(),
         schedule_report_transaction_count_shape => isf_public_interface_schedule_report_transaction_count_shape(),
         schedule_report_transaction_ordering => isf_public_interface_schedule_report_transaction_ordering(),
+        schedule_report_transaction_loop_keys => isf_public_interface_schedule_report_transaction_loop_keys(),
         schedule_report_transaction_wait_keys => isf_public_interface_schedule_report_transaction_wait_keys(),
         schedule_report_transaction_stage_keys => isf_public_interface_schedule_report_transaction_stage_keys(),
         schedule_report_transaction_stage_kind_values => isf_public_interface_schedule_report_transaction_stage_kind_values(),
@@ -394,6 +396,7 @@ sub build_isf_public_interface_contract {
             't/1242-isf-port-binding-conflict-semantics.t',
             't/1243-isf-port-binding-schedule-report.t',
             't/1244-isf-wait-clause-lowering.t',
+            't/1245-isf-transaction-loop-lowering.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',
@@ -511,6 +514,7 @@ sub isf_public_interface_public_top_level_keys {
             schedule_report_transaction_states_shape
             schedule_report_transaction_count_shape
             schedule_report_transaction_ordering
+            schedule_report_transaction_loop_keys
             schedule_report_transaction_wait_keys
             schedule_report_transaction_stage_keys
             schedule_report_transaction_stage_kind_values
@@ -805,6 +809,7 @@ sub isf_public_interface_schedule_report_top_level_keys {
             inferred_storage
             transactions
             transaction_waits
+            transaction_loops
             transaction_stages
             temporal_contracts
             bank_accesses
@@ -1277,6 +1282,22 @@ sub isf_public_interface_schedule_report_transaction_wait_keys {
     ];
 }
 
+sub isf_public_interface_schedule_report_transaction_loop_keys {
+    return [
+        qw(
+            transaction
+            kind
+            condition
+            entry_state
+            decision_states
+            body_start
+            body_states
+            exit_state
+            body_clause_count
+        ),
+    ];
+}
+
 sub isf_public_interface_schedule_report_transaction_stage_keys {
     return [
         qw(
@@ -1380,6 +1401,7 @@ sub isf_public_interface_schedule_report_presence_key_family_map {
         schedule_report_bank_access_keys => isf_public_interface_schedule_report_bank_access_keys(),
         schedule_report_transaction_port_binding_keys => isf_public_interface_schedule_report_transaction_port_binding_keys(),
         schedule_report_transaction_keys => isf_public_interface_schedule_report_transaction_keys(),
+        schedule_report_transaction_loop_keys => isf_public_interface_schedule_report_transaction_loop_keys(),
         schedule_report_transaction_wait_keys => isf_public_interface_schedule_report_transaction_wait_keys(),
         schedule_report_transaction_stage_keys => isf_public_interface_schedule_report_transaction_stage_keys(),
         schedule_report_temporal_contract_keys => isf_public_interface_schedule_report_temporal_contract_keys(),